rewrite to not need job control
when calling this script from some other scripts, bash complained about
not being able to call `fg`.
This could be fixed by activating job control (`set -m`), but that
caused undesirable printing to the shell from bash.

Now the primary program is no longer placed in background but instead
the job that restores the keyboard layout is placed in the background
with a delay in front.